Обновление

+++ dayrui/Fcms/Model/Search.php
@@ -1,9 +1,6 @@
 <?php namespace Phpcmf\Model;
 
-/**
- * http://www.xunruicms.com
- * 本文件是框架系统文件，二次开发时不可以修改本文件，可以通过继承类方法来重写此文件
- **/
+
 
 
 // 模块搜索类
@@ -36,7 +33,7 @@ class Search extends \Phpcmf\Model {
         if (SYS_CACHE_SEARCH) {
             $data = $this->db->table($this->mytable.'_search')->where('id', $id)->get()->getRowArray();
             $time = intval(SYS_CACHE_SEARCH) * 3600;
-            if ($data && $data['inputtime'] < SYS_TIME - $time) {
+            if ($data && $data['inputtime'] + $time < SYS_TIME) {
                 $this->db->table($this->mytable.'_search')->where('id', $id)->delete();
                 $data = [];
             }
@@ -53,7 +50,7 @@ class Search extends \Phpcmf\Model {
             // 主表的字段
             $field = \Phpcmf\Service::L('cache')->get('table-'.SITE_ID, $this->dbprefix($this->mytable));
             if (!$field) {
-                return dr_return_data(0, dr_lang('主表【%s】字段不存在', $this->mytable));
+                return dr_return_data(0, dr_lang('主表[%s]字段不存在', $this->mytable));
             }
 
             $mod_field = $module['field'];
@@ -63,6 +60,10 @@ class Search extends \Phpcmf\Model {
 
             // 默认搜索条件
             $where = [ '`'.$table.'`.`status` = 9' ];
+            /*
+            if (dr_is_app('fstatus') && isset($this->module['field']['fstatus']) && $this->module['field']['fstatus']['ismain']) {
+                $where[] = [ '`'.$table.'`.`fstatus` = 1' ];
+            }*/
 
             // 关键字匹配条件
             if ($param['keyword'] != '') {
@@ -147,38 +148,61 @@ class Search extends \Phpcmf\Model {
 
             // 最大数据量
             $limit = (int)$module['setting']['search']['total'] ? ' LIMIT '.(int)$module['setting']['search']['total'] : '';
-
             // 组合sql查询结果
             $sql = "SELECT `{$table}`.`id` FROM `".$table."` {$where} ORDER BY id ".$limit;
 
-            // 重新生成缓存文件
-            $result = $this->db->query($sql)->getResultArray();
-            if ($result) {
-                $cid = [];
-                // 删除旧数据
-                $this->db->table($this->mytable.'_search')->where('id', $id)->delete();
-                // 入库索引表
-                foreach ($result as $t) {
-                    $cid[] = $t['id'];
+            if ($limit) {
+                // 重新生成缓存文件
+                $result = $this->db->query($sql)->getResultArray();
+                if ($result) {
+                    $cid = [];
+                    // 删除旧数据
+                    $this->db->table($this->mytable.'_search')->where('id', $id)->delete();
+                    // 入库索引表
+                    foreach ($result as $t) {
+                        $cid[] = $t['id'];
+                    }
+                    // 缓存入库
+                    $data = [
+                        'id' => $id,
+                        'catid' => intval($catid),
+                        'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
+                        'keyword' => $param['keyword'] ? $param['keyword'] : '',
+                        'contentid' => @implode(',', $cid),
+                        'inputtime' => SYS_TIME
+                    ];
+                    $this->db->table($this->mytable.'_search')->replace($data);
+                } else {
+                    $data = [
+                        'id' => $id,
+                        'catid' => intval($catid),
+                        'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
+                        'keyword' => $param['keyword'] ? $param['keyword'] : '',
+                        'contentid' => '',
+                    ];
                 }
-                // 缓存入库
-                $data = [
-                    'id' => $id,
-                    'catid' => intval($catid),
-                    'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
-                    'keyword' => $param['keyword'] ? $param['keyword'] : '',
-                    'contentid' => @implode(',', $cid),
-                    'inputtime' => SYS_TIME
-                ];
-                $this->db->table($this->mytable.'_search')->replace($data);
             } else {
-                $data = [
-                    'id' => $id,
-                    'catid' => intval($catid),
-                    'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
-                    'keyword' => $param['keyword'] ? $param['keyword'] : '',
-                    'contentid' => '',
-                ];
+                // 不限搜索数量
+                $ct = $this->db->query("SELECT count(*) as t FROM `".$table."` {$where} ORDER BY id ")->getRowArray();
+                if ($ct['t']) {
+                    $data = [
+                        'id' => $id,
+                        'catid' => intval($catid),
+                        'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
+                        'keyword' => $param['keyword'] ? $param['keyword'] : '',
+                        'contentid' => intval($ct['t']),
+                        'inputtime' => SYS_TIME
+                    ];
+                    $this->db->table($this->mytable.'_search')->replace($data);
+                } else {
+                    $data = [
+                        'id' => $id,
+                        'catid' => intval($catid),
+                        'params' => dr_array2string(['param' => $param, 'sql' => $sql]),
+                        'keyword' => $param['keyword'] ? $param['keyword'] : '',
+                        'contentid' => '',
+                    ];
+                }
             }
         }
